requirements.txt
pystyle~=2.9
PyGithub~=1.59.1